+++ server/src/main/java/sysmon/server/MetricResultToPointProcessor.java
@@ -29,46 +29,18 @@ public class MetricResultToPointProcessor implements Processor {
 
         BatchPoints.Builder batchPoints = BatchPoints
             .database(MetricResultToPointProcessor.influxDbName)
-            .precision(TimeUnit.MILLISECONDS)
+            .precision(TimeUnit.SECONDS)
             .tag("hostname", metricResult.getHostname());
 
         for(Measurement measurement : measurementList) {
 
             Point.Builder point = Point.measurement(metricResult.getName())
-                .time(metricResult.getTimestamp(), TimeUnit.MILLISECONDS)
+                .time(metricResult.getTimestamp(), TimeUnit.SECONDS)
                 .fields(measurement.getFields())
                 .tag(measurement.getTags());
 
-            /*
-            for (Map.Entry<String,String> entry : measurement.getTags().entrySet()) {
-                //log.info("process() - tag: " + entry.getKey() + "=" + entry.getValue());
-                point.tag(entry.getKey(), entry.getValue());
-            }*/
-
-            /*
-            for (Map.Entry<String,Object> entry : measurement.getFields().entrySet()) {
-                //log.info("process() - field: " + entry.getKey() + "=" + entry.getValue());
-                if(entry.getValue() instanceof Float) {
-                    float num = (float) entry.getValue();
-                    point.addField(entry.getKey(), num);
-                } else if(entry.getValue() instanceof Long) {
-                    long num = (long) entry.getValue();
-                    point.addField(entry.getKey(), num);
-                } else if(entry.getValue() instanceof Integer) {
-                    int num = (int) entry.getValue();
-                    point.addField(entry.getKey(), num);
-                } else if(entry.getValue() instanceof Boolean) {
-                    boolean bol = (Boolean) entry.getValue();
-                    point.addField(entry.getKey(), bol);
-                } else if(entry.getValue() instanceof Number) {
-                    Number num = (Number) entry.getValue();
-                    point.addField(entry.getKey(), num);
-                } else {
-                    String str = (String) entry.getValue();
-                    point.addField(entry.getKey(), str);
-                }
-            }*/
             batchPoints.point(point.build());
+
         }
 
         exchange.getIn().setBody(batchPoints.build());